HW4: functions have simplified, a common utility for DNA and RNA has created
