+++ tools/dna_rna_tools.py
@@ -10,6 +10,18 @@ def is_dna(seq: str) -> bool:
     return all(base in "aAtTgGcC" for base in seq)
 
 
+def is_rna(seq: str) -> bool:
+    """
+    A function for checking whether a sequence is a RNA sequence
+
+    Args: seq (str): Input string to check.
+
+    Returns: bool: True if the sequence contains only A, U, G, C.
+
+    """
+    return all(base in "aAuUgGcC" for base in seq)
+
+
 def transcribe(seq: str) -> str:
     """
     A function that transcribes DNA into RNA.
@@ -21,10 +33,27 @@ def transcribe(seq: str) -> str:
     """
     if not is_dna(seq):
         raise ValueError(
-            "This sequence cannot be transcribed because it is not DNA. "
+            "This sequence cannot be transcribed because it is not DNA."
             "For retroviruses, use reverse_transcribe."
         )
     return seq.replace("T", "U").replace("t", "u")
+
+
+def reverse_transcribe(seq: str) -> str:
+    """
+    A function that reverse-transcribes RNA into DNA.
+
+    Args: seq (str): Input string to be reverse transcribed.
+
+    Returns: str: DNA sequence where U are replaced by T.
+
+    """
+    if not is_rna(seq):
+        raise ValueError(
+            "This sequence cannot be transcribed because it is not RNA."
+            "For DNA use transcribe"
+        )
+    return seq.replace("U", "T").replace("u", "t")
 
 
 def reverse(seq: str) -> str:
@@ -48,21 +77,21 @@ def complement(seq: str) -> str:
     Returns: str: Complementary sequence.
 
     """
-    dna = {
-        "A": "T",
-        "T": "A",
-        "G": "C",
-        "C": "G",
-        "a": "t",
-        "t": "a",
-        "g": "c",
-        "c": "g",
+    dna_table = {
+        "A": "T", "T": "A", "G": "C", "C": "G",
+        "a": "t", "t": "a", "g": "c", "c": "g",
+    }
+    rna_table = {
+        "A": "U", "U": "A", "G": "C", "C": "G",
+        "a": "u", "u": "a", "g": "c", "c": "g",
     }
 
     if is_dna(seq):
-        return "".join(dna[base] for base in seq)
+        return "".join(dna_table[base] for base in seq)
+    elif is_rna(seq):
+        return "".join(rna_table[base] for base in seq)
     else:
-        raise ValueError("Sequence should be DNA.")
+        raise ValueError("Sequence should be DNA or RNA.")
 
 
 def reverse_complement(seq: str) -> str:
@@ -74,7 +103,4 @@ def reverse_complement(seq: str) -> str:
     Returns: str: Complement sequence written in reverse order.
 
     """
-    if not is_dna(seq):
-        raise ValueError("Sequence should be DNA.")
-
     return reverse(complement(seq))
